Calculate a mediation analysis for an SEM.

Calculate a mediation analysis for an SEM.

Usage

rmedsem(
  mod,
  indep,
  med,
  dep,
  standardized = FALSE,
  mcreps = NULL,
  approach = "bk",
  p.threshold = 0.05,
  effect.size = c("RIT", "RID")
)

Arguments

mod

A fitted SEM model (lavaan).

indep

A string indicating the name of the independent variable in the model.

med

A string indicating the name of the mediator variable in the model.

dep

A string indicating the name of the dependent variable in the model.

standardized

A boolean indicating whether the coefficients should be standardized. The default value is F.

mcreps

An integer determining the number of monte-carlo samples.

approach

either 'bk' or 'zlc' or both c("bk", "zlc") (default)

p.threshold

A double giving the p-value for determining whether a path is significant or not

effect.size

calculate different effect-sizes; one or more of "RIT", "RID"

Value

A rmedsem structure containing the results from the analysis

Examples

data <- haven::read_stata("https://stats.idre.ucla.edu/stat/data/hsbdemo.dta")
#> Error in open.connection(con, "rb"): Timeout was reached: [stats.idre.ucla.edu] Connection timeout after 10001 ms
mod.txt <- "
read ~ math
science ~ read + math
"
mod <- lavaan::sem(mod.txt, data=data)
#> Error in lavaan::lavaan(model = mod.txt, data = data, model.type = "sem",     int.ov.free = TRUE, int.lv.free = FALSE, auto.fix.first = TRUE,     auto.fix.single = TRUE, auto.var = TRUE, auto.cov.lv.x = TRUE,     auto.cov.y = TRUE, auto.th = TRUE, auto.delta = TRUE, auto.efa = TRUE): lavaan ERROR: data is a function; it should be a data.frame
out <- rmedsem(mod, indep="math", med="read", dep="science",
               standardized=TRUE, mcreps=5000,
               approach = c("bk","zlc"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'nobs': object 'mod' not found
print(out)
#> Error in print(out): object 'out' not found